desloppify: fix type safety and init coupling issues
type safety cluster:
- client.py: replace cast(dict, json.loads()) with _parse_dict() that validates isinstance(value, dict); remove unused cast import
- retry.py: add explicit : Any annotation to RetryableOperation.result
- scientific/validation.py: tighten return type from dict[str, Any] to _ValidationResults TypedDict
- generator.py: int() cast on token count in warmup_model (token counts are integers)

init coupling cluster (all wontfix):
- requests import-time error: hard dep with better error message, not optional
- module-level logger: standard Python idiom, not a side effect
- LOG_LEVEL import-time: standard config loading pattern

+++ infrastructure/llm/review/generator.py
@@ -459,7 +459,7 @@ def warmup_model(client: LLMClient, text_preview: str, model_name: str) -> tuple
         if first_token_time is not None:
             # first_token_time is when the first token arrived; generation_time excludes TTFT
             generation_time = elapsed - (first_token_time - start_time)
-            output_tokens = len(response.split()) * 1.3
+            output_tokens = int(len(response.split()) * 1.3)
             tokens_per_sec = output_tokens / generation_time if generation_time > 0 else 0
         else:
             # No tokens received — streaming produced no output
